% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASPENGLMS-package.R
\docType{package}
\name{ASPENGLMS-package}
\alias{ASPENGLMS}
\alias{ASPENGLMS-package}
\title{ASPENGLMS: beta-binomial mixed models for allelic imbalance}
\description{
Research utilities for fitting and summarising allele-specific expression
models. See the repository tutorial for a synthetic-data workflow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Muquesting/ASPEN_GLMS}
  \item Report bugs at \url{https://github.com/Muquesting/ASPEN_GLMS/issues}
}

}
\author{
\strong{Maintainer}: ASPEN GLMS Team \email{todo@example.com}

Authors:
\itemize{
  \item ASPEN GLMS Team \email{todo@example.com}
}

}
\keyword{internal}
