% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_convert.R
\name{inv_logit}
\alias{inv_logit}
\title{Inverse-logit transform}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{Values on the logit scale.}
}
\value{
Probabilities in (0, 1).
}
\description{
Inverse-logit transform
}
