% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_convert.R
\name{logit}
\alias{logit}
\title{Logit transform with clipping}
\usage{
logit(p, eps = 1e-08)
}
\arguments{
\item{p}{Probabilities in (0, 1).}

\item{eps}{Numerical guard to avoid infinities.}
}
\value{
Transformed values on the logit scale.
}
\description{
Logit transform with clipping
}
