% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_convert.R
\name{logit_from_allelic_ratio}
\alias{logit_from_allelic_ratio}
\title{Convert allelic ratio to logit scale}
\usage{
logit_from_allelic_ratio(ar)
}
\arguments{
\item{ar}{Allelic ratio between 0 and 1.}
}
\value{
Logit-transformed value.
}
\description{
Convert allelic ratio to logit scale
}
